[feature] Add the ability to run commands per group of hosts.
The program will look for the environment attribute `MAX_HOSTS` which
defines the max number of hosts the command will run on in parallel.
If the attribute doesn't exist it will default to the total number of
hosts.

In addition, if the command fails on one or multiple hosts, the program
will exit with the error code 1 instead of 0 (default value).

I also added the binary name to a .gitignore file.